Changed to CacheConfig
