--- ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/config/ApplicationConfig.java
@@ -1,30 +1,0 @@
-package com.nesslabs.nesslabspring.config;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-@RequiredArgsConstructor
-@EnableCaching
-public class ApplicationConfig {
-
-    @Bean
-    public CacheManager cacheManager() {
-        return new ConcurrentMapCacheManager("tags");
-    }
-
-
-}
