+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/config/CacheConfig.java
@@ -1,0 +1,22 @@
+package com.nesslabs.nesslabspring.config;
+
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+@EnableCaching
+public class CacheConfig {
+
+    @Bean
+    public CacheManager cacheManager() {
+        return new ConcurrentMapCacheManager("tags");
+    }
+
+
+}
